+++ src/Plainion.GraphViz/Model/RelaxedGraphBuilder.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Plainion.GraphViz.Model;
 
 namespace Plainion.GraphViz.Model
 {